Move Win32 stream NLS code out of headers
