+++ src/CMakeLists.txt
@@ -49,7 +49,8 @@ set (log4cplus_sources
   threads.cxx
   timehelper.cxx
   tls.cxx
-  version.cxx)
+  version.cxx
+  win32fstream.cxx)
 
 #message (STATUS "Type: ${UNIX}|${CYGWIN}|${WIN32}")
